Split prompt and validator demo into two files
Also add info text when running
